Add $mirror_url parameter to a few common repos
This adds the parameter $mirror_url to the repo classes for CentOS 6,
CentOS 5, Scientific Linx and EPEL.  It's intended to allow the use of
local mirrors through Puppet 3.x databinding.

The parameter takes a "clean" URL to the base of the mirroring
infrastructure for each class, e.g., rsync://msync.centos.org::CentOS
for CentOS.  After checking it agains a regex for clean URLs, it then
interpolates with the known directory structure to create `basedir`
variables for each repo.  The `failovermethod` parameter is set to
`priority` for each repo to avoid round-robin behavior with the existing
`mirrorlist`.